SYMPTOSENSE/requirements.txt
Flask==3.1.2
numpy==2.3.2
pandas==2.3.2
scikit-learn==1.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
